Avalonia - Use MainView Instead of Window
