+++ .github/workflows/android.yml
@@ -1,6 +1,6 @@
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "main", "avalonia" ]
   pull_request:
     branches: [ "main" ]
     types: [ "review_requested", "ready_for_review" ]
