+++ NickvisionApplication.Avalonia/NickvisionApplication.Avalonia/App.axaml.cs
@@ -25,7 +25,7 @@ public partial class App : Application
         }
         else if (ApplicationLifetime is ISingleViewApplicationLifetime singleViewPlatform)
         {
-            singleViewPlatform.MainView = new MainWindow(_mainWindowController, singleViewPlatform).MainContent;
+            singleViewPlatform.MainView = new MainView(_mainWindowController, singleViewPlatform);
         }
         base.OnFrameworkInitializationCompleted();
     }
